updated wavemeter
